削除ボタンと編集ボタンの表示変更

+++ resources/views/photos/detail.blade.php
@@ -4,13 +4,14 @@
         <div class="content-detail">
             <h1 class="detail-title">{{$photo->title}}</h1>
             <img class="detail-image" src="/storage/photos/{{$photo->photo}}" alt="{{$photo->title}}">
-            <p class="detail-description">{{$photo->description}}</p>    
-        @if($photo->user_id === $user->id)
-        <button type=“button” id="detail-link" class="btn btn-link" onclick="location.href='/photos/{{$photo->id}}/edit'">編集する</button>  
-        <button type=“button” id="detail-danger" class="btn btn-danger" onclick="location.href='/photos/{{$photo->id}}/delete'">削除する</button>            
-        @else
-
-        @endif
+            <p class="detail-description">{{$photo->description}}</p>
+            @if(Auth::check())    
+                @if($user->id === $photo->user_id)
+                <button type=“button” id="detail-link" class="btn btn-link" onclick="location.href='/photos/{{$photo->id}}/edit'">編集する</button>  
+                <button type=“button” id="detail-danger" class="btn btn-danger" onclick="location.href='/photos/{{$photo->id}}/delete'">削除する</button>            
+                @endif
+            @else
+            @endif
         </div>
     @else
     <p>画像はありません</p>
